+++ kythe/go/services/xrefs/xrefs_test.go
@@ -17,6 +17,7 @@
 package xrefs
 
 import (
+	"fmt"
 	"regexp"
 	"testing"
 
@@ -54,3 +55,45 @@ func TestFilterRegexp(t *testing.T) {
 		}
 	}
 }
+
+func TestPatcher(t *testing.T) {
+	tests := []struct {
+		oldText, newText string
+
+		oldSpans []*span
+		newSpans []*span
+	}{{
+		oldText:  "this is some text",
+		newText:  "this is some changed text",
+		oldSpans: []*span{{0, 5}, {13, 17}, {8, 17}},
+		newSpans: []*span{{0, 5}, {21, 25}, nil},
+	}, {
+		oldText:  "line one\nline two\nline three\nline four\n",
+		newText:  "line one\nline three\nline two\nline four\n",
+		oldSpans: []*span{{0, 10}, {10, 14}, {16, 19}, {20, 24}, {25, 30}, {29, 38}},
+		newSpans: []*span{{0, 10}, {10, 14}, nil, {22, 26}, nil, {29, 38}},
+	}}
+
+	for _, test := range tests {
+		p := NewPatcher([]byte(test.oldText), []byte(test.newText))
+		if len(test.oldSpans) != len(test.newSpans) {
+			t.Fatalf("Invalid test: {%v}", test)
+		}
+
+		for i, s := range test.oldSpans {
+			start, end, exists := p.Patch(s.start, s.end)
+
+			if ns := test.newSpans[i]; ns == nil && exists {
+				t.Errorf("Expected span not to exist in new text; received (%d, %d]", start, end)
+			} else if ns != nil && !exists {
+				t.Errorf("Expected span %s to exist in new text as %s; did not exist", s, ns)
+			} else if ns != nil && exists && (start != ns.start || end != ns.end) {
+				t.Errorf("Expected %s; received (%d, %d]", ns, start, end)
+			}
+		}
+	}
+}
+
+type span struct{ start, end int32 }
+
+func (s span) String() string { return fmt.Sprintf("(%d, %d]", s.start, s.end) }